+++ trixnity-messenger-compose-app/src/desktopMain/kotlin/de/connect2x/messenger/desktop/MessengerFolderPropertyDefiner.kt
@@ -9,5 +9,5 @@ class MessengerFolderPropertyDefiner : PropertyDefinerBase() {
     override fun getPropertyValue(): String =
         if (System.getenv("TRIXNITY_MESSENGER_ROOT_PATH") == null && BuildConfig.flavor == Flavor.DEV) {
             "./app-data"
-        } else getAppPath(BuildConfig.appNameCleaned).toString()
+        } else getAppPath(BuildConfig.appId).toString()
 }