+++ trixnity-messenger-compose-view/src/androidMain/kotlin/de/connect2x/messenger/compose/view/Util.kt
@@ -1,5 +1,0 @@
-package de.connect2x.messenger.compose.view
-
-fun String.cleanName(): String {
-    return this.replace("[-.\\s]".toRegex(), "").lowercase()
-}