+++ trixnity-messenger/src/jvmMain/kotlin/de/connect2x/trixnity/messenger/util/SecretByteArrayKey.jvm.kt
@@ -5,6 +5,7 @@ import com.sun.jna.Pointer
 import com.sun.jna.platform.win32.Kernel32
 import com.sun.jna.platform.win32.WinDef
 import com.sun.jna.ptr.PointerByReference
+import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.util.*
 import kotlinx.coroutines.Dispatchers
@@ -18,13 +19,14 @@ private val log = KotlinLogging.logger { }
 actual fun platformGetPlatformSecret(): Module = module {
     single<GetPlatformSecret> {
         GetPlatformSecret { id, sizeOnCreate ->
+            val appId = get<MatrixMessengerConfiguration>().appId
             when (getOs()) {
                 OS.MAC_OS, OS.WINDOWS -> {
                     try {
-                        val existingKey = getSecret(id)?.decodeBase64Bytes()
+                        val existingKey = getSecret(appId, id)?.decodeBase64Bytes()
                         if (existingKey == null) {
                             val newKey = SecureRandom.nextBytes(sizeOnCreate)
-                            setSecret(id, newKey.encodeBase64())
+                            setSecret(appId, id, newKey.encodeBase64())
                             newKey
                         } else {
                             existingKey
@@ -41,12 +43,13 @@ actual fun platformGetPlatformSecret(): Module = module {
     }
 }
 
-private suspend fun getSecret(id: String): String? = withContext(Dispatchers.IO) {
+private suspend fun getSecret(appId: String, secretId: String): String? = withContext(Dispatchers.IO) {
     when (getOs()) {
         OS.WINDOWS -> {
+            val targetName = "$appId-$secretId"
             val credentialRef = PointerByReference()
-            val success = WinCredentials.CredReadA(targetName = id, credentialRef = credentialRef)
-            log.debug { "Read secret ('$id') from credentials store: $success" }
+            val success = WinCredentials.CredReadA(targetName = targetName, credentialRef = credentialRef)
+            log.debug { "Read secret ('$targetName') from credentials store: $success" }
             if (success) {
                 try {
                     val credential = Credential(credentialRef.value)
@@ -57,19 +60,20 @@ private suspend fun getSecret(id: String): String? = withContext(Dispatchers.IO)
             } else {
                 when (val error = Kernel32.INSTANCE.GetLastError()) {
                     Kernel32.ERROR_NOT_FOUND -> {
-                        log.warn { "Cannot read secret ('$id') from credentials store. Cannot find target name '$id'" }
+                        log.warn { "Cannot read secret ('$targetName') from credentials store. Cannot find target name '$targetName'" }
                         null
                     }
 
-                    Kernel32.ERROR_NO_SUCH_LOGON_SESSION -> throw SecretNotFoundException("Cannot read secret ('$id') from credentials store. The logon session cannot be found.")
-                    Kernel32.ERROR_INVALID_FLAGS -> throw SecretNotFoundException("Cannot read secret ('$id') from credentials store. The provided flags are invalid.")
-                    else -> throw SecretNotFoundException("Cannot read secret ('$id') from credentials store. Unknown error: $error.")
+                    Kernel32.ERROR_NO_SUCH_LOGON_SESSION -> throw SecretNotFoundException("Cannot read secret ('$targetName') from credentials store. The logon session cannot be found.")
+                    Kernel32.ERROR_INVALID_FLAGS -> throw SecretNotFoundException("Cannot read secret ('$targetName') from credentials store. The provided flags are invalid.")
+                    else -> throw SecretNotFoundException("Cannot read secret ('$targetName') from credentials store. Unknown error: $error.")
                 }
             }
         }
 
         OS.MAC_OS -> {
-            val serviceName = id.toByteArray()
+            val serviceName = appId.toByteArray()
+            val accountName = secretId.toByteArray()
             val passwordSize = IntArray(1)
             val passwordRef = PointerByReference()
             val itemRef = PointerByReference()
@@ -77,15 +81,15 @@ private suspend fun getSecret(id: String): String? = withContext(Dispatchers.IO)
                 null,
                 serviceName.size,
                 serviceName,
-                0,
-                null,
+                accountName.size,
+                accountName,
                 passwordSize,
                 passwordRef,
                 itemRef,
             )
             when (result) {
                 0 -> {
-                    log.debug { "Read secret ('$id') from keychain." }
+                    log.debug { "Read secret ('$secretId') from keychain." }
                     val pointer = passwordRef.value
                     val password = pointer?.getByteArray(0, passwordSize[0])?.decodeToString()
                     MacOsCredentials.SecKeychainItemFreeContent(null, pointer)
@@ -93,13 +97,13 @@ private suspend fun getSecret(id: String): String? = withContext(Dispatchers.IO)
                 }
 
                 -25300 -> {
-                    log.warn { "Cannot find secret ('$id') in keychain." }
+                    log.warn { "Cannot find secret ('$secretId') in keychain." }
                     null
                 }
 
                 else -> {
                     val errorMessage = macOsConvertErrorCodeToMessage(result)
-                    throw SecretNotFoundException("Cannot read secret ('$id') from keychain. ${errorMessage ?: "Unknown error"}")
+                    throw SecretNotFoundException("Cannot read secret ('$secretId') from keychain. ${errorMessage ?: "Unknown error"}")
                 }
             }
         }
@@ -108,52 +112,54 @@ private suspend fun getSecret(id: String): String? = withContext(Dispatchers.IO)
     }
 }
 
-private fun setSecret(id: String, secret: String) {
+private fun setSecret(appId: String, secretId: String, secretValue: String) {
     when (getOs()) {
         OS.WINDOWS -> {
-            val byteArray = secret.toByteArray()
+            val targetName = "$appId-$secretId"
+            val byteArray = secretValue.toByteArray()
             val memory = Memory(byteArray.size.toLong())
             memory.write(0, byteArray, 0, byteArray.size)
             val success = WinCredentials.CredWriteA(
                 Credential(
-                    targetName = id,
+                    targetName = targetName,
                     credentialBlob = memory,
                     credentialBlobSize = byteArray.size
                 ),
                 WinDef.DWORD(0)
             )
             memory.clear()
-            log.debug { "Write secret ('$id') to credentials store: $success" }
+            log.debug { "Write secret ('$targetName') to credentials store: $success" }
             if (success.not()) {
                 when (val error = Kernel32.INSTANCE.GetLastError()) {
-                    Kernel32.ERROR_NO_SUCH_LOGON_SESSION -> log.error { "Cannot write secret ('$id') to credentials store. The logon session cannot be found." }
-                    Kernel32.ERROR_INVALID_PARAMETER -> log.error { "Cannot write secret ('$id') to credentials store. Tried to change protected fields of an existing credential." }
-                    Kernel32.ERROR_INVALID_FLAGS -> log.error { "Cannot write secret ('$id') to credentials store. The provided flags are invalid." }
-                    else -> log.error { "Cannot write secret ('$id') to credentials store. Unknown error: $error." }
+                    Kernel32.ERROR_NO_SUCH_LOGON_SESSION -> log.error { "Cannot write secret ('$targetName') to credentials store. The logon session cannot be found." }
+                    Kernel32.ERROR_INVALID_PARAMETER -> log.error { "Cannot write secret ('$targetName') to credentials store. Tried to change protected fields of an existing credential." }
+                    Kernel32.ERROR_INVALID_FLAGS -> log.error { "Cannot write secret ('$targetName') to credentials store. The provided flags are invalid." }
+                    else -> log.error { "Cannot write secret ('$targetName') to credentials store. Unknown error: $error." }
                 }
             }
         }
 
         OS.MAC_OS -> {
-            val serviceName = id.toByteArray()
+            val serviceName = appId.toByteArray()
+            val accountName = secretId.toByteArray()
             val itemRef = PointerByReference()
-            val secretByteArray = secret.toByteArray()
+            val secretByteArray = secretValue.toByteArray()
             val pointer = itemRef.value
             if (pointer == null) {
                 val result = MacOsCredentials.SecKeychainAddGenericPassword(
                     null,
                     serviceName.size,
                     serviceName,
-                    0,
-                    null,
+                    accountName.size,
+                    accountName,
                     secretByteArray.size,
                     secretByteArray
                 )
                 when (result) {
-                    0 -> log.debug { "Successfully wrote secret ('$id') to keychain." }
+                    0 -> log.debug { "Successfully wrote secret ('$secretId') to keychain." }
                     else -> {
                         val errorMessage = macOsConvertErrorCodeToMessage(result)
-                        log.error { "Cannot save secret ('$id') to keychain. ${errorMessage ?: "Unknown error."}" }
+                        log.error { "Cannot save secret ('$secretId') to keychain. ${errorMessage ?: "Unknown error."}" }
                     }
                 }
             }